# ContentDistributionBackend/src/DACM.ContentDistribution/Models/AssetDistributionMetadata.cs
/*
 * ContentDistributionBackend
 *
 * Service providing content distribution metadata for ordered digital assets
 *
 * OpenAPI spec version: 0.0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace DACM.ContentDistribution.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class AssetDistributionMetadata : IEquatable<AssetDistributionMetadata>
    { 
        /// <summary>
        /// Gets or Sets AssetId
        /// </summary>
        [Required]

        [DataMember(Name="assetId")]
        public Object AssetId { get; set; }

        /// <summary>
        /// Gets or Sets Name
        /// </summary>

        [DataMember(Name="name")]
        public Object Name { get; set; }

        /// <summary>
        /// Gets or Sets FileURL
        /// </summary>

        [DataMember(Name="fileURL")]
        public Object FileURL { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class AssetDistributionMetadata {\n");
            sb.Append("  AssetId: ").Append(AssetId).Append("\n");
            sb.Append("  Name: ").Append(Name).Append("\n");
            sb.Append("  FileURL: ").Append(FileURL).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((AssetDistributionMetadata)obj);
        }

        /// <summary>
        /// Returns true if AssetDistributionMetadata instances are equal
        /// </summary>
        /// <param name="other">Instance of AssetDistributionMetadata to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(AssetDistributionMetadata other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    AssetId == other.AssetId ||
                    AssetId != null &&
                    AssetId.Equals(other.AssetId)
                ) && 
                (
                    Name == other.Name ||
                    Name != null &&
                    Name.Equals(other.Name)
                ) && 
                (
                    FileURL == other.FileURL ||
                    FileURL != null &&
                    FileURL.Equals(other.FileURL)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (AssetId != null)
                    hashCode = hashCode * 59 + AssetId.GetHashCode();
                    if (Name != null)
                    hashCode = hashCode * 59 + Name.GetHashCode();
                    if (FileURL != null)
                    hashCode = hashCode * 59 + FileURL.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(AssetDistributionMetadata left, AssetDistributionMetadata right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(AssetDistributionMetadata left, AssetDistributionMetadata right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}