# ContentDistributionBackend/src/DACM.ContentDistribution/Models/ContentDistributionMetadata.cs
/*
 * ContentDistributionBackend
 *
 * Service providing content distribution metadata for ordered digital assets
 *
 * OpenAPI spec version: 0.0.1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace DACM.ContentDistribution.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class ContentDistributionMetadata : IEquatable<ContentDistributionMetadata>
    { 
        /// <summary>
        /// Gets or Sets DistributionDate
        /// </summary>
        [Required]

        [DataMember(Name="distributionDate")]
        public Object DistributionDate { get; set; }

        /// <summary>
        /// Gets or Sets DistributionChannel
        /// </summary>
        [Required]

        [DataMember(Name="distributionChannel")]
        public Object DistributionChannel { get; set; }

        /// <summary>
        /// Gets or Sets DistributionMethod
        /// </summary>
        [Required]

        [DataMember(Name="distributionMethod")]
        public Object DistributionMethod { get; set; }

        /// <summary>
        /// Gets or Sets Assets
        /// </summary>
        [Required]

        [DataMember(Name="assets")]
        public Object Assets { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class ContentDistributionMetadata {\n");
            sb.Append("  DistributionDate: ").Append(DistributionDate).Append("\n");
            sb.Append("  DistributionChannel: ").Append(DistributionChannel).Append("\n");
            sb.Append("  DistributionMethod: ").Append(DistributionMethod).Append("\n");
            sb.Append("  Assets: ").Append(Assets).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((ContentDistributionMetadata)obj);
        }

        /// <summary>
        /// Returns true if ContentDistributionMetadata instances are equal
        /// </summary>
        /// <param name="other">Instance of ContentDistributionMetadata to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(ContentDistributionMetadata other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    DistributionDate == other.DistributionDate ||
                    DistributionDate != null &&
                    DistributionDate.Equals(other.DistributionDate)
                ) && 
                (
                    DistributionChannel == other.DistributionChannel ||
                    DistributionChannel != null &&
                    DistributionChannel.Equals(other.DistributionChannel)
                ) && 
                (
                    DistributionMethod == other.DistributionMethod ||
                    DistributionMethod != null &&
                    DistributionMethod.Equals(other.DistributionMethod)
                ) && 
                (
                    Assets == other.Assets ||
                    Assets != null &&
                    Assets.Equals(other.Assets)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (DistributionDate != null)
                    hashCode = hashCode * 59 + DistributionDate.GetHashCode();
                    if (DistributionChannel != null)
                    hashCode = hashCode * 59 + DistributionChannel.GetHashCode();
                    if (DistributionMethod != null)
                    hashCode = hashCode * 59 + DistributionMethod.GetHashCode();
                    if (Assets != null)
                    hashCode = hashCode * 59 + Assets.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(ContentDistributionMetadata left, ContentDistributionMetadata right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(ContentDistributionMetadata left, ContentDistributionMetadata right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}